+++ app/build.gradle
@@ -6,8 +6,8 @@ android {
         applicationId "com.wangpm.enteranimation"
         minSdkVersion 19
         targetSdkVersion 33
-        versionCode 1
-        versionName "1.0"
+        versionCode 2
+        versionName "2.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
