+++ app/src/main/java/anim/Anim.java
@@ -1,7 +1,7 @@
 package anim;
 
 import android.graphics.Canvas;
-import android.view.View;
+import android.util.Log;
 
 import com.wangpm.enteranimation.EnterAnimLayout;
 
@@ -13,23 +13,28 @@ public abstract class Anim {
     protected float w;
     protected float h;
     public float totalPaintTime;//控制绘制时间
+    public boolean isExitAnim = false;
 
     public Anim(EnterAnimLayout view) {
         this(view,2000);
     }
 
     public Anim(EnterAnimLayout view, float totalPaintTime) {
+        this(view, totalPaintTime, false);
+    }
+
+    public Anim(EnterAnimLayout view, float totalPaintTime, boolean isExitAnim) {
         this.totalPaintTime = totalPaintTime;
         this.view = view;
         this.view.setAnim(this);
         w = view.getWidth();
         h = view.getHeight();
+        this.isExitAnim = isExitAnim;
     }
 
     public void startAnimation() {
-        view.setmIsAnimaionRun(true);
+        view.setIsAnimationRun(true);
         view.setStartTime(System.currentTimeMillis());
-
         view.invalidate();
     }
     public void startAnimation(long animTime) {
