+++ app/src/main/java/anim/AnimBaiYeChuang.java
@@ -28,7 +28,11 @@ public class AnimBaiYeChuang extends Anim {
             float bottom = top + h / lines * rate;
             path.addRect(0, top, w, bottom, Path.Direction.CW);
         }
-        canvas.clipPath(path);
+        if(isExitAnim) {
+            canvas.clipPath(path, Region.Op.DIFFERENCE);
+        }else{
+            canvas.clipPath(path);
+        }
         canvas.save();
     }
 }
