+++ app/src/main/java/anim/AnimJieTi.java
@@ -44,8 +44,11 @@ public class AnimJieTi extends Anim {
             float bottom = top + h / lineNum;
             path.addRect(left,top,right,bottom, Path.Direction.CW);
         }
-        canvas.clipPath(path);
-
+        if(isExitAnim) {
+            canvas.clipPath(path, Region.Op.DIFFERENCE);
+        }else{
+            canvas.clipPath(path);
+        }
         canvas.save();
     }
 }
