+++ app/src/main/java/anim/AnimPiLie.java
@@ -20,7 +20,11 @@ public class AnimPiLie extends Anim {
     public void handleCanvas(Canvas canvas, float rate) {
         float rectLeft = w / 2 * rate;
         float rectRight = w - w / 2 * rate;
-        canvas.clipRect(rectLeft, 0, rectRight, h, Region.Op.DIFFERENCE);
+        if(isExitAnim) {
+            canvas.clipRect(rectLeft, 0, rectRight, h);
+        }else{
+            canvas.clipRect(rectLeft, 0, rectRight, h, Region.Op.DIFFERENCE);
+        }
         canvas.save();
     }
 }
