+++ app/src/main/java/anim/AnimShanXingZhanKai.java
@@ -3,6 +3,7 @@ package anim;
 import android.graphics.Canvas;
 import android.graphics.Path;
 import android.graphics.RectF;
+import android.graphics.Region;
 import android.view.View;
 
 import com.wangpm.enteranimation.EnterAnimLayout;
@@ -26,7 +27,11 @@ public class AnimShanXingZhanKai extends Anim {
         path1.addArc(oval,270-180*rate,360*rate);
         path1.lineTo(w/2,h/2);
         path1.close();//封闭
-        canvas.clipPath(path1);
+        if(isExitAnim) {
+            canvas.clipPath(path1, Region.Op.DIFFERENCE);
+        }else{
+            canvas.clipPath(path1);
+        }
 
         canvas.save();
     }
